STOCHASTIC PROCESSES AS CURVES IN HILBERT SPACE
by HARALD CRAMÉR

Table of contents
1 Introduction   1
2 Basic Definitions and Auxiliary Concepts   1
3 Projection Operators and Spectral Families   3
4 Distribution Functions and Equivalence Classes   4
5 Spectral Multiplicity Theory   5
6 Prediction Theory   6
7 Correlation Function and Spectral Type   6
8 Existence of Processes with Given Spectral Type   8
Bibliography   13
1  Introduction

In this paper the theory of spectral multiplicity in a separable Hilbert
space will be applied to the study of stochastic processes x(t), where
x(t) is a complex-valued random variable with a finite second-order
moment, while the parameter t may take any real values.
For an account of multiplicity theory one may refer to Chapter 7 of
Stone's book [12] which deals with the case of a separable space. The
treatment of the subject found e.g. in the books by Halmos [6] and
Nakano [10] is mainly concerned with the more general and considerably
more intricate case of a non-separable space.
The considerations will apply to certain classes of curves in a purely
abstract Hilbert space, and it is only a question of terminology when,
throughout this paper, the discussion confines itself to that particular
realization of Hilbert space which has proved useful in probability
theory.
It is finally observed that all statements may be directly generalized
to the case of vector processes of the form (x_1(t),…,x_k(t)).