On the Prediction of Non-Stationary Processes
by N. A. Abdrabbo^a, M. B. Priestley^a

a. University of Manchester
[Received August 1966. Revised January 1967]

Abstract
We consider the problem of linear least-squares prediction for a class
of nonstationary processes which possess evolutionary spectral
representations. It is shown that, under certain conditions, such
processes admit moving-average representations in terms of
time-dependent coefficients. This feature enables us to develop a close
analogue of the Wiener-Kolmogorov approach to the corresponding problem
for stationary processes.
Table of contents
1 Introduction   1
2 Evolutionary Spectral Representations   3
3 Moving-average Representations for Oscillatory Processes   4
3.1 Continuous-parameter Processes   4
4 The Time-domain Approach   5
4.1 Continuous-parameter Processes   5
5 The Frequency-domain Approach   6
6 Discussion   8
References   9
Bibliography   9