The Thaana character ޥ comes from this line in the Parser.java code:
```java
(char) 0x107A5
```

This appears twice:
1. In `lowercaseSuperscriptAlphabet` array (position for 'q')
2. In `superscripts` array

The issue: **Unicode codepoint U+107A5 is intended to represent
superscript 'q', but Java's `char` type is only 16 bits.** When casting
`0x107A5` to `char`, it truncates to 16 bits:

```
0x107A5 & 0xFFFF = 0x07A5
```

So instead of getting superscript q (U+107A5, which requires surrogate
pairs in Java), the code produces U+07A5, which is ޥ (Thaana Letter
Waavu).

**Fix:** Remove or replace `(char) 0x107A5` since superscript q cannot
be represented as a single `char` in Java. Either use a String with
surrogate pairs or omit superscript q entirely if not needed.