arb.measure.MeasureSpace<X, Σ extends σField<? extends X>, μ extends Measure<? extends X, ? extends Σ>>
A MeasureSpace is a basic Space of MeasureTheory which generalizes the
notion of volume. It contains an underlying set X, the measurable
subsets constituting the σField and the method that is used for
measuring (the Measure). One important example of a measure space is a
ProbabilitySpace. A measurable space consists of the first two
components without a specific measure, thus. A measure space is a triple
(X,Σ,μ) where

 X is a set
 Σ is the σField of the set X
 μ()  is a Measure on the MeasurableSpace (X,Σ)

In other words, a measure space consists of a MeasurableSpace (X,Σ )
together with a Measure on it and thus MeasureSpace extends
MeasurableSpace and provides the method μ() to specify the measure